Middleware support validate: Update tests, move example class from package to app #52
