+++ app/Http/routes.php
@@ -15,26 +15,6 @@ Route::get('/', function () {
     return view('welcome');
 });
 
-Route::post('/posts', [
-    'middleware' => 'permission:create-post',
-    function () {
-        return response()->json(null, 200);
-    }
-]);
-
-Route::group(['middleware'=>'routePermission'], function() {
-
-    Route::post('/blog/{id}', function ($id) {
-        return response()->json(null, 200);
-    });
-});
-
-Route::post('/blog', ['middleware'=>'validate:\PhpSoft\Users\Validators\BlogValidate',
-    function () {
-        return response()->json(null, 200);
-    }
-]);
-
 Route::post('/auth/login', '\PhpSoft\Users\Controllers\AuthController@login');
 Route::group(['middleware'=>'jwt.auth'], function() {
