+++ tests/TestCase.php
@@ -20,6 +20,26 @@ class TestCase extends Illuminate\Foundation\Testing\TestCase
 
         $app->make(Illuminate\Contracts\Console\Kernel::class)->bootstrap();
 
+        Route::post('/posts', [
+            'middleware' => 'permission:create-post',
+            function () {
+                return response()->json(null, 200);
+            }
+        ]);
+
+        Route::group(['middleware'=>'routePermission'], function() {
+
+            Route::post('/blog/{id}', function ($id) {
+                return response()->json(null, 200);
+            });
+        });
+
+        Route::post('/user', ['middleware'=>'validate:App\Http\Validators\UserValidate',
+            function () {
+                return response()->json(null, 200);
+            }
+        ]);
+
         return $app;
     }
